python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.19",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "boto3-stubs[s3]>=1.34.0",
    "mypy>=1.8.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
